lib: refacto move Buffers to SCLang and refacto WIP
