+++ renardo_lib/renardo_lib/SCLang/__init__.py
@@ -1,5 +1,7 @@
 from renardo_lib.SCLang.SCLangExperimentalBindings import *
-from renardo_lib.SCLang.SynthDef import SynthDefs, SynthDef, SynthDefProxy, SampleSynthDef
+from renardo_lib.SCLang.SynthDef import SynthDefs, SynthDef
+from renardo_lib.SCLang.SynthDefProxy import SynthDefProxy
+from renardo_lib.SCLang.SampleSynthDefs import SampleSynthDef
 # from renardo_lib.SCLang import Env
